CIV-16579 Refactor hearing estimate text to new class
