+++ src/test/java/uk/gov/hmcts/reform/civil/service/docmosis/hearingorder/HearingOrderGeneratorTest.java
@@ -27,6 +27,7 @@ import uk.gov.hmcts.reform.civil.model.genapplication.GAJudgesHearingListGAspec;
 import uk.gov.hmcts.reform.civil.model.genapplication.GAOrderCourtOwnInitiativeGAspec;
 import uk.gov.hmcts.reform.civil.model.genapplication.GAOrderWithoutNoticeGAspec;
 import uk.gov.hmcts.reform.civil.sampledata.CaseDataBuilder;
+import uk.gov.hmcts.reform.civil.service.JudicialTimeEstimateHelper;
 import uk.gov.hmcts.reform.civil.service.docmosis.DocmosisService;
 import uk.gov.hmcts.reform.civil.service.docmosis.DocumentGeneratorService;
 import uk.gov.hmcts.reform.civil.service.documentmanagement.SecuredDocumentManagementService;
@@ -67,6 +68,8 @@ class HearingOrderGeneratorTest {
     private HearingOrderGenerator hearingOrderGenerator;
     @MockBean
     private DocmosisService docmosisService;
+    @MockBean
+    private JudicialTimeEstimateHelper timeEstimateHelper;
 
     @Test
     void shouldGenerateHearingOrderDocument() {
@@ -135,6 +138,7 @@ class HearingOrderGeneratorTest {
         void whenJudgeMakeDecision_ShouldGetHearingOrderData() {
             when(docmosisService.getCaseManagementLocationVenueName(any(), any()))
                 .thenReturn(LocationRefData.builder().epimmsId("2").externalShortName("Reading").build());
+            when(timeEstimateHelper.getEstimatedHearingLength(any())).thenReturn("15 minutes");
             CaseData caseData = CaseDataBuilder.builder()
                 .hearingOrderApplication(YesOrNo.NO, YesOrNo.YES).build().toBuilder()
                 .isMultiParty(YES)
@@ -190,6 +194,7 @@ class HearingOrderGeneratorTest {
         void whenJudgeMakeDecision_ShouldGetHearingOrderData() {
             when(docmosisService.getCaseManagementLocationVenueName(any(), any()))
                 .thenReturn(LocationRefData.builder().epimmsId("2").externalShortName("Reading").build());
+            when(timeEstimateHelper.getEstimatedHearingLength(any())).thenReturn("15 minutes");
             CaseData caseData = CaseDataBuilder.builder()
                 .hearingOrderApplication(YesOrNo.NO, YesOrNo.YES).build().toBuilder()
                 .isMultiParty(YES)
@@ -232,6 +237,7 @@ class HearingOrderGeneratorTest {
         void whenJudgeMakeDecision_ShouldGetHearingOrderData_Option2() {
             when(docmosisService.getCaseManagementLocationVenueName(any(), any()))
                 .thenReturn(LocationRefData.builder().epimmsId("2").externalShortName("Manchester").build());
+            when(timeEstimateHelper.getEstimatedHearingLength(any())).thenReturn("15 minutes");
             CaseData caseData = CaseDataBuilder.builder()
                 .hearingOrderApplication(YesOrNo.NO, YesOrNo.YES).build().toBuilder()
                 .isMultiParty(NO)
@@ -286,6 +292,7 @@ class HearingOrderGeneratorTest {
         void whenJudgeMakeDecision_ShouldGetHearingOrderData_Option3() {
             when(docmosisService.getCaseManagementLocationVenueName(any(), any()))
                 .thenReturn(LocationRefData.builder().epimmsId("2").externalShortName("London").build());
+            when(timeEstimateHelper.getEstimatedHearingLength(any())).thenReturn("15 minutes");
             CaseData caseData = CaseDataBuilder.builder()
                 .hearingOrderApplication(YesOrNo.NO, YesOrNo.YES).build().toBuilder()
                 .isMultiParty(NO)
@@ -331,6 +338,7 @@ class HearingOrderGeneratorTest {
         void whenJudgeMakeDecision_ShouldGetHearingOrderData_Option3_1v1() {
             when(docmosisService.getCaseManagementLocationVenueName(any(), any()))
                 .thenReturn(LocationRefData.builder().epimmsId("2").externalShortName("Reading").build());
+            when(timeEstimateHelper.getEstimatedHearingLength(any())).thenReturn("15 minutes");
             CaseData caseData = CaseDataBuilder.builder()
                 .hearingOrderApplication(YesOrNo.NO, YesOrNo.YES).build().toBuilder()
                 .isMultiParty(YES)
@@ -371,144 +379,6 @@ class HearingOrderGeneratorTest {
                 () -> assertEquals(templateData.getJudgeRecital(), caseData.getJudicialGeneralHearingOrderRecital()),
                 () -> assertEquals(templateData.getHearingOrder(), caseData.getJudicialGOHearingDirections())
             );
-        }
-
-        @Test
-        void whenJudgeSpecifiesOtherHearingDuration1_ShouldGetHearingOrderData() {
-            when(docmosisService.getCaseManagementLocationVenueName(any(), any()))
-                .thenReturn(LocationRefData.builder().epimmsId("2").externalShortName("Reading").build());
-            CaseData caseData = CaseDataBuilder.builder()
-                .hearingOrderApplication(YesOrNo.NO, YesOrNo.YES).build().toBuilder()
-                .isMultiParty(YES)
-                .build();
-            caseData = caseData.toBuilder()
-                .judicialListForHearing(caseData.getJudicialListForHearing().toBuilder()
-                                            .judicialTimeEstimate(GAHearingDuration.OTHER)
-                                            .judicialTimeEstimateDays("2")
-                                            .judicialTimeEstimateHours("2")
-                                            .judicialTimeEstimateMinutes("30").build())
-                .build();
-
-            var templateData = hearingOrderGenerator.getTemplateData(null, caseData, "auth", FlowFlag.ONE_RESPONDENT_REPRESENTATIVE);
-
-            assertEquals("2 days, 2 hours and 30 minutes", templateData.getEstimatedHearingLength());
-        }
-
-        @Test
-        void whenJudgeSpecifiesOtherHearingDuration2_ShouldGetHearingOrderData() {
-            when(docmosisService.getCaseManagementLocationVenueName(any(), any()))
-                .thenReturn(LocationRefData.builder().epimmsId("2").externalShortName("Reading").build());
-            CaseData caseData = CaseDataBuilder.builder()
-                .hearingOrderApplication(YesOrNo.NO, YesOrNo.YES).build().toBuilder()
-                .isMultiParty(YES)
-                .build();
-            caseData = caseData.toBuilder()
-                .judicialListForHearing(caseData.getJudicialListForHearing().toBuilder()
-                                            .judicialTimeEstimate(GAHearingDuration.OTHER)
-                                            .judicialTimeEstimateDays("1")
-                                            .judicialTimeEstimateMinutes("30").build())
-                .build();
-
-            var templateData = hearingOrderGenerator.getTemplateData(null, caseData, "auth", FlowFlag.ONE_RESPONDENT_REPRESENTATIVE);
-
-            assertEquals("1 day and 30 minutes", templateData.getEstimatedHearingLength());
-        }
-
-        @Test
-        void whenJudgeSpecifiesOtherHearingDuration3_ShouldGetHearingOrderData() {
-            when(docmosisService.getCaseManagementLocationVenueName(any(), any()))
-                .thenReturn(LocationRefData.builder().epimmsId("2").externalShortName("Reading").build());
-            CaseData caseData = CaseDataBuilder.builder()
-                .hearingOrderApplication(YesOrNo.NO, YesOrNo.YES).build().toBuilder()
-                .isMultiParty(YES)
-                .build();
-            caseData = caseData.toBuilder()
-                .judicialListForHearing(caseData.getJudicialListForHearing().toBuilder()
-                                            .judicialTimeEstimate(GAHearingDuration.OTHER)
-                                            .judicialTimeEstimateDays("2")
-                                            .judicialTimeEstimateHours("6").build())
-                .build();
-
-            var templateData = hearingOrderGenerator.getTemplateData(null, caseData, "auth", FlowFlag.ONE_RESPONDENT_REPRESENTATIVE);
-
-            assertEquals("2 days and 6 hours", templateData.getEstimatedHearingLength());
-        }
-
-        @Test
-        void whenJudgeSpecifiesOtherHearingDuration4_ShouldGetHearingOrderData() {
-            when(docmosisService.getCaseManagementLocationVenueName(any(), any()))
-                .thenReturn(LocationRefData.builder().epimmsId("2").externalShortName("Reading").build());
-            CaseData caseData = CaseDataBuilder.builder()
-                .hearingOrderApplication(YesOrNo.NO, YesOrNo.YES).build().toBuilder()
-                .isMultiParty(YES)
-                .build();
-            caseData = caseData.toBuilder()
-                .judicialListForHearing(caseData.getJudicialListForHearing().toBuilder()
-                                            .judicialTimeEstimate(GAHearingDuration.OTHER)
-                                            .judicialTimeEstimateHours("2")
-                                            .judicialTimeEstimateMinutes("30").build())
-                .build();
-
-            var templateData = hearingOrderGenerator.getTemplateData(null, caseData, "auth", FlowFlag.ONE_RESPONDENT_REPRESENTATIVE);
-
-            assertEquals("2 hours and 30 minutes", templateData.getEstimatedHearingLength());
-        }
-
-        @Test
-        void whenJudgeSpecifiesOtherHearingDuration5_ShouldGetHearingOrderData() {
-            when(docmosisService.getCaseManagementLocationVenueName(any(), any()))
-                .thenReturn(LocationRefData.builder().epimmsId("2").externalShortName("Reading").build());
-            CaseData caseData = CaseDataBuilder.builder()
-                .hearingOrderApplication(YesOrNo.NO, YesOrNo.YES).build().toBuilder()
-                .isMultiParty(YES)
-                .build();
-            caseData = caseData.toBuilder()
-                .judicialListForHearing(caseData.getJudicialListForHearing().toBuilder()
-                                            .judicialTimeEstimate(GAHearingDuration.OTHER)
-                                            .judicialTimeEstimateHours("6").build())
-                .build();
-
-            var templateData = hearingOrderGenerator.getTemplateData(null, caseData, "auth", FlowFlag.ONE_RESPONDENT_REPRESENTATIVE);
-
-            assertEquals("6 hours", templateData.getEstimatedHearingLength());
-        }
-
-        @Test
-        void whenJudgeSpecifiesOtherHearingDuration6_ShouldGetHearingOrderData() {
-            when(docmosisService.getCaseManagementLocationVenueName(any(), any()))
-                .thenReturn(LocationRefData.builder().epimmsId("2").externalShortName("Reading").build());
-            CaseData caseData = CaseDataBuilder.builder()
-                .hearingOrderApplication(YesOrNo.NO, YesOrNo.YES).build().toBuilder()
-                .isMultiParty(YES)
-                .build();
-            caseData = caseData.toBuilder()
-                .judicialListForHearing(caseData.getJudicialListForHearing().toBuilder()
-                                            .judicialTimeEstimate(GAHearingDuration.OTHER)
-                                            .judicialTimeEstimateDays("3").build())
-                .build();
-
-            var templateData = hearingOrderGenerator.getTemplateData(null, caseData, "auth", FlowFlag.ONE_RESPONDENT_REPRESENTATIVE);
-
-            assertEquals("3 days", templateData.getEstimatedHearingLength());
-        }
-
-        @Test
-        void whenJudgeSpecifiesOtherHearingDuration7_ShouldGetHearingOrderData() {
-            when(docmosisService.getCaseManagementLocationVenueName(any(), any()))
-                .thenReturn(LocationRefData.builder().epimmsId("2").externalShortName("Reading").build());
-            CaseData caseData = CaseDataBuilder.builder()
-                .hearingOrderApplication(YesOrNo.NO, YesOrNo.YES).build().toBuilder()
-                .isMultiParty(YES)
-                .build();
-            caseData = caseData.toBuilder()
-                .judicialListForHearing(caseData.getJudicialListForHearing().toBuilder()
-                                            .judicialTimeEstimate(GAHearingDuration.OTHER)
-                                            .judicialTimeEstimateMinutes("45").build())
-                .build();
-
-            var templateData = hearingOrderGenerator.getTemplateData(null, caseData, "auth", FlowFlag.ONE_RESPONDENT_REPRESENTATIVE);
-
-            assertEquals("45 minutes", templateData.getEstimatedHearingLength());
         }
     }
 }
